Added a colour for the viewer date.
Added a style for the reset button on the Post it toolbar.

Created a border for the profile pic on the viewer.

Fixed the design of the borders on the post it and log cache listview. Also added a message to indicate to users when there aren't any logs.

Added commands for the right-click menu which appears on log cache items.

Added the reset toolbutton's path to the Paths.xaml.

Added a style for the author text on the viewer page.