Modified AuthService and implemented sign in and sign out. It works with sqlite instead of Firebase now.
Overhauled the Cachemaster, DATAMASTER, DATAWRITER, DATAREADER, etc so that it can work with Data Logger 1.3. Code has been refactored and improved and is fully compatible with the new UI. Bugs still need to be checked and the class still has implementations outstanding.

Added helper methods for DATAMASTER classes DataService. Also created an interface for DATAMASTER related methods so they can be accessed from DataService.

Created an "interface" for Cachemaster class as well so it can be accessed from DataService.

Modified NavigationService to help implement the Editor functionality.

Added startup functionality for the log cache.

Modified behaviour of the navigation arrows.

Added edit functionality for PostIt editor.