Removed unecessary usings.
Refactored code for readability and added functionality to keep track of the currently active post it field.

Updated UIs to keep up with the improved user controls. Tooltips also added. Changed some fields to combo boxes in the LoggerCreatePage for cases where user input should be limited. Changed the StackPanel housing the postit list and toolbar to a Grid.